Recognize 1.#INF etc - support can be specified by parameter
When using the Visual Studio compiler, text representations
1.#INF, 1.#IND, 1.#QNAN and 1.#SNAN are used instead of 'inf' and
'NaN'. This patch recognizes them if a parameter for fread
(R version) is set to TRUE.

In the C code, the functions Strtod and readfile were moved to
fread_readfile_template.h. Some functions were renamed, e.g.
Strtod was renamed to TEMPLATE_Strtod. By setting defines in
fread.c and including fread_readfile_template.h, two versions
of the readfile function are created:

- one uses the default code
- the other uses strtod_wrapper and strtold_wrapper, which add
extra checks to handle e.g. 1.#INF and 1.#IND

Depending on the flag vs.inf.nan in the R fread function, one of
the two C functions is called.

This way, the original behaviour is still the default, and runs
without any performance penalty. If needed, the user can activate
the slightly slower modified code which performs extra checks.
