+++ man/fread.Rd
@@ -17,7 +17,8 @@ dec=if (sep!=".") "." else ",", col.names,
 check.names=FALSE, encoding="unknown", quote="\"", 
 strip.white=TRUE, fill=FALSE, blank.lines.skip=FALSE, key=NULL, 
 showProgress=getOption("datatable.showProgress"),   # default: TRUE
-data.table=getOption("datatable.fread.datatable")   # default: TRUE
+data.table=getOption("datatable.fread.datatable"),   # default: TRUE
+vs.inf.nan=FALSE
 )
 }
 \arguments{
@@ -47,6 +48,7 @@ data.table=getOption("datatable.fread.datatable")   # default: TRUE
   \item{key}{Character vector of one or more column names which is passed to \code{\link{setkey}}. It may be a single comma separated string such as \code{key="x,y,z"}, or a vector of names such as \code{key=c("x","y","z")}. Only valid when argument \code{data.table=TRUE}.}
   \item{showProgress}{ TRUE displays progress on the console using \code{\\r}. It is produced in fread's C code where the very nice (but R level) txtProgressBar and tkProgressBar are not easily available. }
   \item{data.table}{ TRUE returns a \code{data.table}. FALSE returns a \code{data.frame}. }
+  \item{vs.inf.nan}{ If set to TRUE, functions are used (with a small performance penalty) that interpret e.g. "1.#INF" and "1.#IND" as the numbers infinity and non-a-number (NaN). }
 }
 \details{
